server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["kjpar0317 <kjpar_kr@yahoo.co.kr>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
uvicorn = "^0.17.5"
pydantic = "^1.9.0"
SQLAlchemy = "^1.4.32"
Jinja2 = "^3.0.3"
mysqlclient = "^2.1.0"
cryptography = "^36.0.1"
python-dotenv = "^0.19.2"
email-validator = "^1.1.3"
PyJWT = "^2.3.0"
gunicorn = "^20.1.0"
uuid = "^1.30"
asyncio = "^3.4.3"
aiomysql = "^0.1.1"
autopep8 = "^1.6.0"
sqlmodel = "^0.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
